ci(host): fail the build on a clippy warning
Nothing linted Rust here, so "no linter configured" and "no lint
failures" looked identical from outside the repo.

`-D warnings` is the whole point of the step. clippy exits 0 with its
warnings on stdout, so without it the job is green whether the code is
clean or not, and a lint nobody fails is a lint nobody fixes.
--all-targets covers the test builds too; the tray's WM_COMMAND cast was
only reported there.

Four warnings existed and three were mechanical: an unused Painter
import, a usize cast to usize, and a literal passed through an empty
format string.

The fourth is a finding rather than noise. Manifest parses `name`,
`description` and `hotkeys` and reads none of them. They stay, because
they are the manifest contract rather than leftovers, and the allow is
scoped to that struct so the next field that stops being read for a worse
reason still gets reported. Worth knowing that mods/timetracker binds
three hotkeys the host has never once fired.

+++ host/src/main.rs
@@ -16,7 +16,7 @@ mod surface;
 mod tray;
 
 use aether_raster::{Backend, Font};
-use aether_runtime::{Driver, Painter, RasterPainter, Rgb};
+use aether_runtime::{Driver, RasterPainter, Rgb};
 use aether_window::{Button, Event, Window};
 use std::collections::HashMap;
 use std::path::PathBuf;
@@ -259,7 +259,7 @@ fn run() -> Result<(), String> {
                 painted_frames += 1;
             }
             if last_report.elapsed() >= Duration::from_secs(1) {
-                let n = painted_frames.max(1) as u32;
+                let n = painted_frames.max(1);
                 println!(
                     "[dew] {frames} fps | painted {painted_frames} | solve {:?} | raster+blit {:?}",
                     sum_frame / n,
